finished most of logic for ending the round
only thing left to do to end the round is change the result of the if statements and and make a screen for round over. After that I  will be testing. Im also working on the logic for a team getting points for answering correctly. Next goal is to create propts for the user to follow thourghtout the game and place everything where it needs to go?